tox: disable proxies when running nosetests
httpretty does not handle proxy well, but the need for http_proxy and
https_proxy is present if you're needing to install pip packages
behind a proxy.

Right now this will fail:
  http_proxy=http://my.proxy/ tox --recreate

Example failures are in
   UrlHelperWaitForUrlsTest.test_url_wait_for

it seems reasonable that we would never want proxy set when running tests
as we should never be doing real http traffic.

This just wraps our calls to nosetests with something that removes
the http_proxy friends.

Change-Id: I96585acaed0c8d70e1925b3859c85c87d1e07b2f