fix 'tox -e docs' by limiting sphinx versions
This bug is reported in http://pad.lv/1379998

this will fix it, but admittedly will break again if sphinx
package is released with a version that is not blacklisted here.

This will have to do until a sufficent fix is available in pbr.
Then the hard restrictions on sphinx could be removed.

Change-Id: I696fc3e4ef8803bbb2a95e34501d5330a64ab93c